requirements.txt
kafka-python
opencv-python==4.8.0.74
Flask